requirements.txt
Flask>=2.1
pandas>=1.5
numpy>=1.24
yfinance>=0.2.27
python-dotenv>=1.0
google-generativeai>=0.5.2
crewai>=0.1.0
litellm>=0.2.0
requests
python-dotenv

